asdf

+++ allocator.c
@@ -56,7 +56,7 @@ void * malloc(size_t size){
         
         int i = log(v)/log(2) - 3;
         int* page_start = map_list[i];
-        int* next_page = page_start + 1;
+        long* next_page = page_start + 1;
         while(*next_page != (long)NULL){
             page_start = (int*)*next_page;
             next_page = page_start + 1;
